requirements.txt
streamlit
pandas
gspread
plotly
google-auth